remove legacy versions
