+++ conan/tools/__init__.py
@@ -1,2 +1,5 @@
+from conans.model.version import Version
 CONAN_TOOLCHAIN_ARGS_FILE = "conanbuild.conf"
 CONAN_TOOLCHAIN_ARGS_SECTION = "toolchain"
+
+
